+++ .idea/workspace.xml
@@ -4,9 +4,16 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="e10ed917-6337-413a-b38c-015e3c6c4951" name="Changes" comment="gitignore">
+    <list default="true" id="e10ed917-6337-413a-b38c-015e3c6c4951" name="Changes" comment="tweaks to admin">
+      <change afterPath="$PROJECT_DIR$/TurnPageRoot/profiles/urls.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/TurnPageRoot/TurnPageRoot/settings.py" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/TurnPageRoot/settings.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/TurnPageRoot/TurnPageRoot/urls.py" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/TurnPageRoot/urls.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/admin.py" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/admin.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/templates/bookSwiping/login.html" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/profiles/templates/profiles/login.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/templates/bookSwiping/signup.html" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/profiles/templates/profiles/signup.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/urls.py" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/urls.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/views.py" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/bookSwiping/views.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/TurnPageRoot/templates/navbar.html" beforeDir="false" afterPath="$PROJECT_DIR$/TurnPageRoot/templates/navbar.html" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -52,6 +59,7 @@
 }</component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/TurnPageRoot/profiles/templates/profiles" />
       <recent name="$PROJECT_DIR$/TurnPageRoot/static/images" />
     </key>
   </component>
@@ -67,7 +75,7 @@
       <workItem from="1665155065417" duration="77000" />
       <workItem from="1666211024272" duration="253000" />
       <workItem from="1666270999100" duration="2011000" />
-      <workItem from="1666298829754" duration="867000" />
+      <workItem from="1666298829754" duration="1261000" />
     </task>
     <task id="LOCAL-00001" summary="boilerplate setup">
       <created>1665150211698</created>
@@ -118,7 +126,14 @@
       <option name="project" value="LOCAL" />
       <updated>1666299005182</updated>
     </task>
-    <option name="localTasksCounter" value="8" />
+    <task id="LOCAL-00008" summary="tweaks to admin">
+      <created>1666299715619</created>
+      <option name="number" value="00008" />
+      <option name="presentableId" value="LOCAL-00008" />
+      <option name="project" value="LOCAL" />
+      <updated>1666299715619</updated>
+    </task>
+    <option name="localTasksCounter" value="9" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -142,6 +157,7 @@
     <MESSAGE value="small updates" />
     <MESSAGE value="merged" />
     <MESSAGE value="gitignore" />
-    <option name="LAST_COMMIT_MESSAGE" value="gitignore" />
+    <MESSAGE value="tweaks to admin" />
+    <option name="LAST_COMMIT_MESSAGE" value="tweaks to admin" />
   </component>
 </project>
