+++ TurnPageRoot/TurnPageRoot/urls.py
@@ -19,4 +19,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('bookSwiping.urls')),
+    path('accounts/', include('profiles.urls')),
 ]
