+++ TurnPageRoot/bookSwiping/admin.py
@@ -22,4 +22,4 @@ admin.site.register(Genre)
 admin.site.register(BookGenre)
 admin.site.register(UserGenre)
 admin.site.register(Bookshelf)
-admin.site.register(User)
+# admin.site.register(User)
