+++ TurnPageRoot/bookSwiping/urls.py
@@ -3,6 +3,6 @@ from . import views
 
 urlpatterns = [
     path('', views.HomeView.as_view(), name='home'),
-    path('login', views.LoginView.as_view(), name='login'),
-    path('signup', views.SignupView.as_view(), name='signup')
+    # path('login', views.LoginView.as_view(), name='login'),
+    # path('signup', views.SignupView.as_view(), name='signup')
 ]
