+++ TurnPageRoot/profiles/urls.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('login', views.LoginView.as_view(), name='login'),
+    path('signup', views.SignupView.as_view(), name='signup')
+]
